Make necessary changes to bring our code standard up to Airbnb's ESLint preset.
