+++ src/index.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import RegistrationPage from './pages/RegistrationPage';
+
+ReactDOM.render(<RegistrationPage />, document.getElementById('root'));
